Fix zero receive window advertisements.
With the recent changes db36d948fa63ce950d94a5e8e9ebc37956543661, we try
to balance the receive window advertisements between payload lengths vs
segment overhead length. This works fine when segment size are much
higher than the overhead, but not otherwise. In cases where the segment
length is smaller than the segment overhead, we may end up not
advertising zero receive window for long time and end up tail-dropping
segments. This is especially pronounced when application socket reads
are slow or stopped. In this change we advertise zero receive window and
shrink the previously advertised window if we do not have any room to
enqueue newer segments.

Fixes #4903

PiperOrigin-RevId: 342273080